Potentially Finished Draft
**Original number of rows: 330 / New: 277**
<br>
- For 1 of the entries, a yearsOpen value of 2019 was found. Makes the whole entry likely invalid. Row was dropped.
- The rest of the rows were dropped due to both AdultWeekend and AdultWeekday price being absent This is important for ultimately deciding what the price of tickets should be

**Original number of columns: 27 / New: 25**
<br>
- fastEight column was dropped due to most values being 0
- AdultWeekday was dropped due to AdultWeekend having more price data

**Silverton Mountain has valid skiable terrain**
<br>
After finding a strange distribution for skiable terrain, we looked further and found Silverton has an unusually high value. After a google search, we found out this may have been a mistake when the data was being entered. The correct data was put in.

<br>
Misc:
<br>
State data regarding population and square mileage was saved. This could be useful for determining ticket pricing later.